requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
pydantic>=2.7.0
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-ollama>=0.1.0
requests>=2.32.0

